[#47] refactor: 타입정의 type.ts 에서 types 폴더로 이동
