+++ src/components/postDetailPage/BackToListBtn.tsx
@@ -1,6 +1,6 @@
+import { Post } from '@/types/post';
 import { Button } from '../ui/button';
 import { Link } from 'react-router-dom';
-import { Post } from '@/lib/types';
 
 type BackToListBtnProps = {
   data: Post;
