+++ src/components/postDetailPage/ContentSection.tsx
@@ -1,4 +1,4 @@
-import { Post } from '@/lib/types';
+import { Post } from '@/types/post';
 
 type ContentSectionProps = {
   data: Post;
