+++ src/components/postDetailPage/UserInfo.tsx
@@ -1,4 +1,4 @@
-import { Post } from '@/lib/types';
+import { Post } from '@/types/post';
 
 type UserInfoProps = {
   data: Post;
