+++ src/components/postListPage/SortOptions.tsx
@@ -1,5 +1,5 @@
 import { SORT_OPTIONS } from '@/lib/data';
-import { OptionType } from '@/lib/types';
+import { OptionType } from '@/types/post';
 import { cn } from '@/lib/utils';
 import { DotFilledIcon } from '@radix-ui/react-icons';
