+++ src/mocks/mockData/post/allPosts.ts
@@ -1,4 +1,4 @@
-import { Post } from '@/lib/types';
+import { Post } from '@/types/post';
 
 export const allPosts: Post[] = [
   // DAILY
